+++ internal/acctest/acctest.go
@@ -253,6 +253,69 @@ data "davinci_connection" "crowd_strike" {
 //
 // The `resourceName` input can be a random charset and will be used for the name of
 // each resource and datasource in the returned hcl.
+// p1services is a list of services besides SSO and DaVinci to enable on the environment
+func PingoneEnvrionmentServicesSsoHcl(resourceName string, p1Services []string) (hcl string) {
+	adminEnvID := os.Getenv("PINGONE_ENVIRONMENT_ID")
+	licenseID := os.Getenv("PINGONE_LICENSE_ID")
+	username := os.Getenv("PINGONE_USERNAME")
+	// loop through p1services for adding to hcl
+	p1ServicesHcl := ""
+	for _, s := range p1Services {
+		p1ServicesHcl = p1ServicesHcl + fmt.Sprintf(`
+	service {
+		type = "%s"
+	}
+	`, s)
+	}
+
+	return fmt.Sprintf(`
+resource "pingone_environment" "%[1]s" {
+	name = "tf-testacc-dv-dynamic-%[1]s"
+	type = "SANDBOX"
+	license_id = "%[2]s"
+  default_population {
+	}
+	service {
+		type = "SSO"
+	}
+	service {
+		type = "DaVinci"
+	}
+	%[5]s
+}
+
+data "pingone_role" "%[1]s" {
+	name = "DaVinci Admin"
+}
+
+data "pingone_user" "%[1]s"{
+	username             = "%[3]s"
+	environment_id       = "%[4]s"
+}
+
+resource "pingone_role_assignment_user" "%[1]s" {
+	environment_id       = "%[4]s"
+	user_id              = data.pingone_user.%[1]s.id
+	role_id              = data.pingone_role.%[1]s.id
+	scope_environment_id = resource.pingone_environment.%[1]s.id
+}
+
+data "davinci_connections" "read_all" {
+	environment_id = resource.pingone_role_assignment_user.%[1]s.scope_environment_id
+}
+
+`, resourceName, licenseID, username, adminEnvID, p1ServicesHcl)
+}
+
+// PingoneEnvrionmentSsoHcl returns hcl for a pingone environment and assigns roles used for SSO by davinci
+// The following environment vars must be set:
+// - PINGONE_ENVIRONMENT_ID
+// - PINGONE_LICENSE_ID
+// - PINGONE_SSO_USERNAME
+// - PINGONE_SSO_PASSWORD
+//
+// The `resourceName` input can be a random charset and will be used for the name of
+// each resource and datasource in the returned hcl.
 func PingoneEnvrionmentSsoHcl(resourceName string) (hcl string) {
 	adminEnvID := os.Getenv("PINGONE_ENVIRONMENT_ID")
 	licenseID := os.Getenv("PINGONE_LICENSE_ID")