+++ internal/service/davinci/resource_application.go
@@ -3,12 +3,14 @@ package davinci
 import (
 	"context"
 	"fmt"
+	"regexp"
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/pingidentity/terraform-provider-davinci/internal/sdk"
+	"github.com/pingidentity/terraform-provider-davinci/internal/utils"
 	dv "github.com/samir-gandhi/davinci-client-go/davinci"
 )
 
@@ -382,7 +384,7 @@ func ResourceApplication() *schema.Resource {
 			},
 		},
 		Importer: &schema.ResourceImporter{
-			StateContext: schema.ImportStatePassthroughContext,
+			StateContext: resourceApplicationImport,
 		},
 	}
 }
@@ -578,6 +580,36 @@ func resourceApplicationDelete(ctx context.Context, d *schema.ResourceData, meta
 	d.SetId("")
 
 	return diags
+}
+
+func resourceApplicationImport(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+
+	idComponents := []utils.ImportComponent{
+		{
+			Label:  "environment_id",
+			Regexp: regexp.MustCompile(`[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}`),
+		},
+		{
+			Label:     "davinci_application_id",
+			Regexp:    regexp.MustCompile(`[a-f0-9]{32}`),
+			PrimaryID: true,
+		},
+	}
+
+	attributes, err := utils.ParseImportID(d.Id(), idComponents...)
+	if err != nil {
+		return nil, err
+	}
+
+	if err = d.Set("environment_id", attributes["environment_id"]); err != nil {
+		return nil, err
+	}
+
+	d.SetId(attributes["davinci_application_id"])
+
+	resourceApplicationRead(ctx, d, meta)
+
+	return []*schema.ResourceData{d}, nil
 }
 
 func expandApp(d *schema.ResourceData) (*dv.AppUpdate, error) {